pyproject.toml
[project]
name = "data-agent"
version = "0.0.1"
dependencies = [
    "python-dotenv", 
    "psycopg2-binary", 
    "google-adk", 
    "chainlit",
    "fastapi",
    "uvicorn",
    "requests",
    "prefect"
]

[tool.setuptools]
py-modules = ["agents", "app", "db"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
